Safari-611.1.10/LayoutTests/media/modern-media-controls/compact-media-controls/compact-media-controls-constructor-expected.txt

Testing the CompactMediaControls constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.element.localName is "div"
PASS mediaControls.element.className is "compact media-controls"
PASS mediaControls.layoutTraits is LayoutTraits.Compact
PASS mediaControls.state is CompactMediaControls.States.Paused

PASS successfullyParsed is true

TEST COMPLETE
